Рефакторит модуль utils и подключает вьюху отствия точек маршрута
