+++ src/presenter/points-presenter.js
@@ -2,6 +2,7 @@ import ListPointsView from '../view/list-points-view.js';
 import SortView from '../view/sort-view.js';
 import EditPointView from '../view/edit-point-view.js';
 import PointView from '../view/point-view.js';
+import EmptyPointView from '../view/empty-point-view.js';
 import {render, replace} from '../framework/render.js';
 
 
@@ -23,12 +24,8 @@ export default class PointsPresenter {
     this.#points = [...this.#pointsModel.points];
     this.#offers = this.#pointsModel.offers;
     this.#destinations = this.#pointsModel.destinations;
-    render(new SortView(), this.#pointsContainer);
-    render(this.#listPointsComponent, this.#pointsContainer);
 
-    for (let i = 0; i < this.#points.length; i++) {
-      this.#renderPoint(this.#points[i]);
-    }
+    this.#renderPointsBoard();
   }
 
   #getOffersByType(type) {
@@ -37,6 +34,9 @@ export default class PointsPresenter {
 
   #getOffersById(type, itemsId) {
     const offersType = this.#getOffersByType(type);
+    if (!offersType) {
+      return [];
+    }
     return offersType.offers.filter((offer) => itemsId.find((id) => offer.id === id));
   }
 
@@ -84,5 +84,18 @@ export default class PointsPresenter {
 
     render(pointComponent, this.#listPointsComponent.element);
 
+  }
+
+  #renderPointsBoard() {
+    if (this.#points.length === 0) {
+      return render(new EmptyPointView, this.#pointsContainer);
+    }
+
+    render(new SortView(), this.#pointsContainer);
+    render(this.#listPointsComponent, this.#pointsContainer);
+
+    for (let i = 0; i < this.#points.length; i++) {
+      this.#renderPoint(this.#points[i]);
+    }
   }
 }
