+++ src/view/empty-point-view.js
@@ -1,0 +1,11 @@
+import AbstractView from '../framework/view/abstract-view.js';
+
+function createEmptyPointTemplate() {
+  return '<p class="trip-events__msg">Click New Event button to create your first point</p>';
+}
+
+export default class EmptyPointView extends AbstractView {
+  get template() {
+    return createEmptyPointTemplate();
+  }
+}
